// src/usecase/customer/update/update.customer.dto.ts
export interface InputUpdateCustomerDto {
    id: string,
    name: string,
    address: {
        street: string;
        city: string;
        number: number;
        zip: string;
    }
}

export interface OutputUpdateCustomerDto {
    id: string,
    name: string,
    address: {
        street: string;
        city: string;
        number: number;
        zip: string;
    }
}